# improve short-circuiting on (any|drop)null(!)