motivation
This PR implements logic tests for `anynull` to avoid checking arrays that cannot contain `Nullable` elements. This PR supersedes https://github.com/JuliaStats/NullableArrays.jl/pull/188 by also fixing `dropnull` and `dropnull!` by extending the same logic tests to account for `Vector{Union}`s, which are handled incorrectly by the current `dropnull` and `dropnull!`

benchmarks and tests
```julia
using NullableArrays, BenchmarkTools, Base.Test
srand(1); x = rand(1_000_000); @benchmark anynull(x)
srand(1); x = rand(1000, 1000); @benchmark anynull(x)
srand(1); x = convert(Array{Union{Int, Float64}}, rand(1_000_000)); @benchmark anynull(x)
srand(1); x = convert(Array{Union{Int, Float64}}, rand(1000, 1000)); @benchmark anynull(x)
x = convert(Array{Union{Int, Nullable{Int}}}, repmat([1, Nullable()], 5)); dropnull!(x)
x = convert(Array{Union{Int, Nullable{Int}}}, repmat([1, Nullable()], 5)); dropnull(x)
x = convert(Array{Union{Int, Nullable{Int}, Float64, Nullable{Float64}}}, repmat([1, Nullable()], 5)); dropnull!(x)
x = convert(Array{Union{Int, Nullable{Int}, Float64, Nullable{Float64}}}, repmat([1, Nullable()], 5)); dropnull(x)
```

results
````julia
# benchmark
srand(1); x = rand(1_000_000); @benchmark anynull(x)
# PR
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     17.377 ns (0.00% GC)
  median time:      17.669 ns (0.00% GC)
  mean time:        19.026 ns (0.00% GC)
  maximum time:     81.627 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     998
  time tolerance:   5.00%
  memory tolerance: 1.00%
# master
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     270.970 μs (0.00% GC)
  median time:      283.878 μs (0.00% GC)
  mean time:        318.548 μs (0.00% GC)
  maximum time:     1.278 ms (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%


############################################################################################
# benchmark
srand(1); x = rand(1000, 1000); @benchmark anynull(x)
# PR
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     16.856 ns (0.00% GC)
  median time:      17.410 ns (0.00% GC)
  mean time:        18.608 ns (0.00% GC)
  maximum time:     94.667 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     998
  time tolerance:   5.00%
  memory tolerance: 1.00%
# master
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     270.965 μs (0.00% GC)
  median time:      286.472 μs (0.00% GC)
  mean time:        311.942 μs (0.00% GC)
  maximum time:     1.024 ms (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%


############################################################################################
# benchmark
srand(1); x = convert(Array{Union{Int, Float64}}, rand(1_000_000)); @benchmark anynull(x)
# PR
BenchmarkTools.Trial:
  memory estimate:  128 bytes
  allocs estimate:  2
  --------------
  minimum time:     443.242 ns (0.00% GC)
  median time:      488.109 ns (0.00% GC)
  mean time:        513.604 ns (1.91% GC)
  maximum time:     10.368 μs (92.00% GC)
  --------------
  samples:          10000
  evals/sample:     198
  time tolerance:   5.00%
  memory tolerance: 1.00%
# master
BenchmarkTools.Trial:
  memory estimate:  15.26 MiB
  allocs estimate:  1000000
  --------------
  minimum time:     31.832 ms (0.00% GC)
  median time:      34.032 ms (1.36% GC)
  mean time:        34.662 ms (1.02% GC)
  maximum time:     42.589 ms (0.00% GC)
  --------------
  samples:          144
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%

############################################################################################
# benchmark
srand(1); x = convert(Array{Union{Int, Float64}}, rand(1000, 1000)); @benchmark anynull(x)
# PR
BenchmarkTools.Trial:
  memory estimate:  128 bytes
  allocs estimate:  2
  --------------
  minimum time:     445.313 ns (0.00% GC)
  median time:      481.639 ns (0.00% GC)
  mean time:        514.056 ns (1.98% GC)
  maximum time:     11.591 μs (90.53% GC)
  --------------
  samples:          10000
  evals/sample:     198
  time tolerance:   5.00%
  memory tolerance: 1.00%
# master
BenchmarkTools.Trial:
  memory estimate:  15.26 MiB
  allocs estimate:  1000000
  --------------
  minimum time:     12.168 ms (0.00% GC)
  median time:      14.276 ms (4.17% GC)
  mean time:        14.866 ms (3.17% GC)
  maximum time:     28.467 ms (3.35% GC)
  --------------
  samples:          335
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%


############################################################################################
# test
x = convert(Array{Union{Int, Nullable{Int}}}, repmat([1, Nullable()], 5)); dropnull!(x)
# PR
5-element Array{Int64,1}:
 1
 1
 1
 1
 1
# master
10-element Array{Union{Int64, Nullable{Int64}},1}:
 1
 #NULL
 1
 #NULL
 1
 #NULL
 1
 #NULL
 1
 #NULL

############################################################################################
# test
x = convert(Array{Union{Int, Nullable{Int}}}, repmat([1, Nullable()], 5)); dropnull(x)
# PR
5-element Array{Int64,1}:
 1
 1
 1
 1
 1
# master
10-element Array{Union{Int64, Nullable{Int64}},1}:
 1
 #NULL
 1
 #NULL
 1
 #NULL
 1
 #NULL
 1
 #NULL

############################################################################################
# test
x = convert(Array{Union{Int, Nullable{Int}, Float64, Nullable{Float64}}}, repmat([1, Nullable()], 5)); dropnull!(x)
# PR
5-element Array{Union{Float64, Int64},1}:
 1
 1
 1
 1
 1
# master
10-element Array{Union{Float64, Int64, Nullable{Float64}, Nullable{Int64}},1}:
 1
 #NULL
 1
 #NULL
 1
 #NULL
 1
 #NULL
 1
 #NULL


############################################################################################
# test
x = convert(Array{Union{Int, Nullable{Int}, Float64, Nullable{Float64}}}, repmat([1, Nullable()], 5)); dropnull(x)
# PR
5-element Array{Union{Float64, Int64},1}:
 1
 1
 1
 1
 1
# master
10-element Array{Union{Float64, Int64, Nullable{Float64}, Nullable{Int64}},1}:
 1
 #NULL
 1
 #NULL
 1
 #NULL
 1
 #NULL
 1
 #NULL
```