@cjprybol 2017-04-02
That'd be a great improvement! Would you like to hold off on this PR and see how things play out with that issue? I could remove the anynull changes and leave the dropnull fixes.