@nalimilan 2017-04-15
Actually, @mbauman just gave me a much simpler way to check whether a type can hold a `Nullable`: `typeintersect(T, Nullable) !== Union{}`. So we can simplify `_isnullable` a lot.