@ararslan 2017-04-15
Something we did in DataArrays that you might want to consider here is deprecating `allnull(x)` and `anynull(x)` in favor of `all(isnull, x)` and `any(isnull, x)`, respectively. The same level of specialization can still occur; you just overload e.g. `all(::typeof(isnull), x::NullableArray)`.