@cjprybol 2017-04-15
I like both of those suggestions, closing in favor of https://github.com/JuliaStats/NullableArrays.jl/pull/192. I'll use mbauman's much better `_isnullable` in the linked DataTables PR. Thanks all!